Add the PermissionService

+++ src/controllers/user.controller.mjs
@@ -1,8 +1,11 @@
 import UserService from '../services/user.service.js';
+import PermissionService from '../services/permission.service.js';
+import NodeError from '../models/node-error.js';
 
 class UserController {
   constructor() {
     this.userService = new UserService();
+    this.permissionService = new PermissionService();
   }
 
   searchUser = async (req, res, next) => {
@@ -17,6 +20,11 @@ class UserController {
   getUserById = async (req, res, next) => {
     try {
       const userId = req.params.id;
+
+      if (!this.permissionService.isUserCanRequestUserData(userId)) {
+        throw new NodeError(403, 'You have no rights to request this data');
+      }
+
       const user = await this.userService.getUserById(userId);
       res.send(user);
     } catch (error) {
@@ -35,6 +43,12 @@ class UserController {
   
   updateUser = async (req, res, next) => {
     try {
+      const userId = req.body?.id || req.body?._id;
+
+      if (!this.permissionService.isUserCanRequestUserData(userId)) {
+        throw new NodeError(403, 'You have no rights to request this data');
+      }
+
       const user = await this.userService.updateUser(req.body);
       res.send(user);
     } catch (error) {
@@ -45,6 +59,11 @@ class UserController {
   deleteUser = async (req, res, next) => {
     try {
       const userId = req.params.id;
+
+      if (!this.permissionService.isUserCanRequestUserData(userId)) {
+        throw new NodeError(403, 'You have no rights to request this data');
+      }
+
       const deleteResult = await this.userService.deleteUser(userId);
       res.send(deleteResult);
     } catch (error) {
